requirements.txt
python-telegram-bot
python-dotenv
websockets
pandas
numpy